Ex24 & 25: Review of what we've covered so far.
These two focus on adding everything we've covered so far and
separate out the string and variables from the functions. Ex24
has the string, printing, variable information, while ex25
focuses on the specific use of functions.

Functions are created in ex25.py then are imported into the
python interpreter for use.