ex36: changing format of story.
Now the passageway leads to a raft you can get on to continue your journey.
You are thrown from the raft again and again. The rafts pile up each journey through.
You have to get a special item and return to the original raft to continue your way
instead of being stuck in this Groundhog Day nightmare of a journey.: